[Inductor] Fix rand_like with kwargs device of str type
